requirements.txt
streamlit
youtube-transcript-api
groq
yt-dlp
openai-whisper
